# swagger.json
{
          "openapi": "3.1.0",
          "info": {
                    "title": "FastAPI Boilerplate",
                    "description": "FastAPI Boilerplate",
                    "version": "1.0.0"
          },
          "paths": {
...
                    "/api/v1/users/sign-up": {
                              "post": {
                                        "tags": [
                                                  "Authentication"
                                        ],
                                        "summary": "Sign Up",
                                        "operationId": "sign_up_api_v1_users_sign_up_post",
                                        "requestBody": {
                                                  "content": {
                                                            "application/json": {
                                                                      "schema": {
                                                                                "$ref": "#/components/schemas/CreateUserRequest"
                                                                      }
                                                            }
                                                  },
                                                  "required": true
                                        },
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/User"
                                                                                }
                                                                      }
                                                            }
                                                  },
                                                  "422": {
                                                            "description": "Validation Error",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/HTTPValidationError"
                                                                                }
                                                                      }
                                                            }
                                                  }
                                        }
                              }
                    },
                    "/api/v1/users/login": {
                              "post": {
                                        "tags": [
                                                  "Authentication"
                                        ],
                                        "summary": "Login",
                                        "description": "Generate JWT token using email and password",
                                        "operationId": "login_api_v1_users_login_post",
                                        "requestBody": {
                                                  "content": {
                                                            "application/json": {
                                                                      "schema": {
                                                                                "$ref": "#/components/schemas/LoginRequest"
                                                                      }
                                                            }
                                                  },
                                                  "required": true
                                        },
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {}
                                                                      }
                                                            }
                                                  },
                                                  "422": {
                                                            "description": "Validation Error",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/HTTPValidationError"
                                                                                }
                                                                      }
                                                            }
                                                  }
                                        }
                              }
                    },
                    "/api/v1/users/update/password": {
                              "post": {
                                        "tags": [
                                                  "Authentication"
                                        ],
                                        "summary": "Update Password",
                                        "operationId": "update_password_api_v1_users_update_password_post",
                                        "requestBody": {
                                                  "content": {
                                                            "application/json": {
                                                                      "schema": {
                                                                                "$ref": "#/components/schemas/UpdatePasswordRequest"
                                                                      }
                                                            }
                                                  },
                                                  "required": true
                                        },
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {}
                                                                      }
                                                            }
                                                  },
                                                  "422": {
                                                            "description": "Validation Error",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/HTTPValidationError"
                                                                                }
                                                                      }
                                                            }
                                                  }
                                        },
                                        "security": [
                                                  {
                                                            "HTTPBearer": []
                                                  }
                                        ]
                              }
                    },
                    "/api/v1/users/logout": {
                              "get": {
                                        "tags": [
                                                  "Authentication"
                                        ],
                                        "summary": "Logout",
                                        "description": "invalidates the jwt token",
                                        "operationId": "logout_api_v1_users_logout_get",
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {}
                                                                      }
                                                            }
                                                  }
                                        },
                                        "security": [
                                                  {
                                                            "HTTPBearer": []
                                                  }
                                        ]
                              }
                    },
                    "/api/v1/users/search": {
                              "get": {
                                        "tags": [
                                                  "Users"
                                        ],
                                        "summary": "Search Users",
                                        "description": "Search users by name for email",
                                        "operationId": "search_users_api_v1_users_search_get",
                                        "security": [
                                                  {
                                                            "HTTPBearer": []
                                                  }
                                        ],
                                        "parameters": [
                                                  {
                                                            "name": "q",
                                                            "in": "query",
                                                            "required": false,
                                                            "schema": {
                                                                      "anyOf": [
                                                                                {
                                                                                          "type": "string"
                                                                                },
                                                                                {
                                                                                          "type": "null"
                                                                                }
                                                                      ],
                                                                      "default": "",
                                                                      "title": "Q"
                                                            }
                                                  },
                                                  {
                                                            "name": "page_size",
                                                            "in": "query",
                                                            "required": false,
                                                            "schema": {
                                                                      "anyOf": [
                                                                                {
                                                                                          "type": "integer"
                                                                                },
                                                                                {
                                                                                          "type": "null"
                                                                                }
                                                                      ],
                                                                      "default": 10,
                                                                      "title": "Page Size"
                                                            }
                                                  },
                                                  {
                                                            "name": "page_num",
                                                            "in": "query",
                                                            "required": false,
                                                            "schema": {
                                                                      "anyOf": [
                                                                                {
                                                                                          "type": "integer"
                                                                                },
                                                                                {
                                                                                          "type": "null"
                                                                                }
                                                                      ],
                                                                      "default": 1,
                                                                      "title": "Page Num"
                                                            }
                                                  },
                                                  {
                                                            "name": "sort_dir",
                                                            "in": "query",
                                                            "required": false,
                                                            "schema": {
                                                                      "anyOf": [
                                                                                {
                                                                                          "type": "string"
                                                                                },
                                                                                {
                                                                                          "type": "null"
                                                                                }
                                                                      ],
                                                                      "default": "asc",
                                                                      "title": "Sort Dir"
                                                            }
                                                  },
                                                  {
                                                            "name": "sort_column",
                                                            "in": "query",
                                                            "required": false,
                                                            "schema": {
                                                                      "anyOf": [
                                                                                {
                                                                                          "type": "string"
                                                                                },
                                                                                {
                                                                                          "type": "null"
                                                                                }
                                                                      ],
                                                                      "default": "id",
                                                                      "title": "Sort Column"
                                                            }
                                                  }
                                        ],
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {}
                                                                      }
                                                            }
                                                  },
                                                  "422": {
                                                            "description": "Validation Error",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/HTTPValidationError"
                                                                                }
                                                                      }
                                                            }
                                                  }
                                        }
                              }
                    },
                    "/api/v1/users/profile/me": {
                              "get": {
                                        "tags": [
                                                  "Users"
                                        ],
                                        "summary": "Get Self User Profile",
                                        "operationId": "get_self_user_profile_api_v1_users_profile_me_get",
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/SelfUserProfileResponse"
                                                                                }
                                                                      }
                                                            }
                                                  }
                                        },
                                        "security": [
                                                  {
                                                            "HTTPBearer": []
                                                  }
                                        ]
                              }
                    },
                    "/api/v1/users/profile/{user_id}": {
                              "get": {
                                        "tags": [
                                                  "User Profile"
                                        ],
                                        "summary": "Get User Profile",
                                        "operationId": "get_user_profile_api_v1_users_profile__user_id__get",
                                        "security": [
                                                  {
                                                            "HTTPBearer": []
                                                  }
                                        ],
                                        "parameters": [
                                                  {
                                                            "name": "user_id",
                                                            "in": "path",
                                                            "required": true,
                                                            "schema": {
                                                                      "type": "integer",
                                                                      "title": "User Id"
                                                            }
                                                  }
                                        ],
                                        "responses": {
                                                  "200": {
                                                            "description": "Successful Response",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/UserProfileResponse"
                                                                                }
                                                                      }
                                                            }
                                                  },
                                                  "422": {
                                                            "description": "Validation Error",
                                                            "content": {
                                                                      "application/json": {
                                                                                "schema": {
                                                                                          "$ref": "#/components/schemas/HTTPValidationError"
                                                                                }
                                                                      }
                                                            }
                                                  }
                                        }
                              }
                    },
...
          },
          "components": {
                    "schemas": {
...
                              "HTTPValidationError": {
                                        "properties": {
                                                  "detail": {
                                                            "items": {
                                                                      "$ref": "#/components/schemas/ValidationError"
                                                            },
                                                            "type": "array",
                                                            "title": "Detail"
                                                  }
                                        },
                                        "type": "object",
                                        "title": "HTTPValidationError"
                              },
                              "LoginRequest": {
                                        "properties": {
                                                  "email": {
                                                            "type": "string",
                                                            "title": "Email",
                                                            "description": "email of the user",
                                                            "examples": [
                                                                      "user@test.com"
                                                            ]
                                                  },
                                                  "password": {
                                                            "type": "string",
                                                            "title": "Password",
                                                            "description": "Encrypted password of the user",
                                                            "examples": [
                                                                      "5f4dcc3b5aa765d61d8327deb882cf99"
                                                            ]
                                                  }
                                        },
                                        "type": "object",
                                        "required": [
                                                  "email",
                                                  "password"
                                        ],
                                        "title": "LoginRequest"
                              },
...
                    },
                    "securitySchemes": {
                              "HTTPBearer": {
                                        "type": "http",
                                        "scheme": "bearer"
                              }
                    }
          }
}